python/requirements.txt
python-uinput
pySerial
